requirements.txt
fastapi
slowapi
redis[asyncio]
pydantic
ray[serve]
sqlalchemy
pika
uvicorn
aiohttp
uuid
jsonschema
logging
requests
aio_pika
asyncio
docker
argparse
python-dotenv
starlette
vllm
pyyaml
functools
os
prometheus_client
prometheus-fastapi-instrumentator
docker.io 
docker-compose